Test for updating repositories with the same dependency (#2158)

+++ src/Microsoft.DotNet.Darc/tests/Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests/VmrTestsBase.cs
@@ -26,40 +26,40 @@ namespace Microsoft.DotNet.Darc.Tests.VirtualMonoRepo;
 
 public abstract class VmrTestsBase
 {
+    protected LocalPath CurrentTestDirectory { get; private set; } = null!;
+    protected LocalPath ProductRepoPath { get; private set; } = null!;
+    protected LocalPath VmrPath { get; private set; } = null!;
+    protected LocalPath TmpPath { get; private set; } = null!;
+    protected LocalPath SecondRepoPath { get; private set; } = null!;
+    protected LocalPath DependencyRepoPath { get; private set; } = null!;
+    protected LocalPath InstallerRepoPath { get; private set; } = null!;
+    protected GitOperationsHelper GitOperations { get; } = new();
+    protected VmrInfo VmrInfo { get; private set; } = null!;
+    
     private Lazy<IServiceProvider> _serviceProvider = null!;
     private readonly CancellationTokenSource _cancellationToken = new();
-    protected LocalPath _currentTestDirectory = null!;
-    protected LocalPath _privateRepoPath = null!;
-    protected LocalPath _vmrPath = null!;
-    protected LocalPath _tmpPath = null!;
-    protected LocalPath _externalRepoPath = null!;
-    protected LocalPath _dependencyRepoPath = null!;
-    protected LocalPath _specialRepoPath = null!;
-    protected LocalPath _installerRepoPath = null!;
-    protected GitOperationsHelper GitOperations { get; } = new();
-    protected VmrInfo vmrInfo = null!;
-   
+
     [SetUp]
     public async Task Setup()
     {
         var testsDirName = "_tests";
-        _currentTestDirectory = VmrTestsOneTimeSetUp.TestsDirectory / testsDirName / Path.GetRandomFileName();
-        Directory.CreateDirectory(_currentTestDirectory);
+        CurrentTestDirectory = VmrTestsOneTimeSetUp.TestsDirectory / testsDirName / Path.GetRandomFileName();
+        Directory.CreateDirectory(CurrentTestDirectory);
 
-        _tmpPath = _currentTestDirectory / "tmp";
-        _privateRepoPath = _currentTestDirectory / "test-repo";
-        _vmrPath = _currentTestDirectory / "vmr";
-        _externalRepoPath = _currentTestDirectory / "external-repo";
-        _dependencyRepoPath = _currentTestDirectory / "dependency";
-        _specialRepoPath = _currentTestDirectory / "special-repo";
-        _installerRepoPath = _currentTestDirectory / "installer";
+        TmpPath = CurrentTestDirectory / "tmp";
+        ProductRepoPath = CurrentTestDirectory / Constants.ProductRepoName;
+        VmrPath = CurrentTestDirectory / "vmr";
+        SecondRepoPath = CurrentTestDirectory / Constants.SecondRepoName;
+        DependencyRepoPath = CurrentTestDirectory / Constants.DependencyRepoName;
+        InstallerRepoPath = CurrentTestDirectory / Constants.InstallerRepoName;
 
-        Directory.CreateDirectory(_tmpPath);
-
-        await CopyVmrForCurrentTest();
+        Directory.CreateDirectory(TmpPath);
+        
         await CopyReposForCurrentTest();
+        await CopyVmrForCurrentTest();
+        
         _serviceProvider = new(CreateServiceProvider);
-        vmrInfo = (VmrInfo)_serviceProvider.Value.GetRequiredService<IVmrInfo>();
+        VmrInfo = (VmrInfo)_serviceProvider.Value.GetRequiredService<IVmrInfo>();
     }
 
     [TearDown]
@@ -67,9 +67,9 @@ public abstract class VmrTestsBase
     {
         try
         {
-            if (_currentTestDirectory is not null)
+            if (CurrentTestDirectory is not null)
             {
-                VmrTestsOneTimeSetUp.DeleteDirectory(_currentTestDirectory.ToString());
+                VmrTestsOneTimeSetUp.DeleteDirectory(CurrentTestDirectory.ToString());
             }
         }
         catch
@@ -88,8 +88,8 @@ public abstract class VmrTestsBase
         .AddVmrManagers(
         sp => sp.GetRequiredService<GitFileManagerFactory>(),
         "git",
-        _vmrPath,
-        _tmpPath,
+        VmrPath,
+        TmpPath,
         null,
         null)
         .BuildServiceProvider();
@@ -102,7 +102,7 @@ public abstract class VmrTestsBase
         var expectedFiles = new List<LocalPath>
         {
             vmrPath / VmrInfo.GitInfoSourcesDir / AllVersionsPropsFile.FileName,
-            vmrInfo.GetSourceManifestPath(),
+            VmrInfo.GetSourceManifestPath(),
             vmrPath / VmrInfo.SourcesDir / VmrInfo.SourceMappingsFileName
         };
 
@@ -209,28 +209,32 @@ public abstract class VmrTestsBase
     internal async Task<string> CopyRepoAndCreateVersionDetails(
         LocalPath currentTestDir,
         string repoName,
-        IDictionary<string, List<Dependency>>? dependencies = null)
+        IDictionary<string, List<string>>? dependencies = null)
     {
         var repoPath = currentTestDir / repoName;
-        CopyDirectory(VmrTestsOneTimeSetUp.TestsDirectory / repoName, repoPath);
         
         var dependenciesString = new StringBuilder();
         if (dependencies != null && dependencies.ContainsKey(repoName))
         {
             var repoDependencies = dependencies[repoName];
-            foreach (var dep in repoDependencies)
+            foreach (var dependencyName in repoDependencies)
             {
-                string sha = await CopyRepoAndCreateVersionDetails(currentTestDir, dep.Name, dependencies);
+                string sha = await CopyRepoAndCreateVersionDetails(currentTestDir, dependencyName, dependencies);
                 dependenciesString.AppendLine(
                     string.Format(
                         Constants.DependencyTemplate,
-                        new[] { dep.Name, dep.Uri, sha }));
+                        new[] { dependencyName, currentTestDir/ dependencyName, sha }));
             }
         }
 
-        var versionDetails = string.Format(Constants.VersionDetailsTemplate, dependenciesString);
-        File.WriteAllText(repoPath / VersionFiles.VersionDetailsXml, versionDetails);
-        await GitOperations.CommitAll(repoPath, "update version details");
+        if (!Directory.Exists(repoPath))
+        {
+            CopyDirectory(VmrTestsOneTimeSetUp.TestsDirectory / repoName, repoPath);
+            var versionDetails = string.Format(Constants.VersionDetailsTemplate, dependenciesString);
+            File.WriteAllText(repoPath / VersionFiles.VersionDetailsXml, versionDetails);
+            await GitOperations.CommitAll(repoPath, "update version details");
+        }
+
         return await GitOperations.GetRepoLastCommit(repoPath);
     }
 
@@ -243,9 +247,9 @@ public abstract class VmrTestsBase
             DefaultIgnoreCondition = System.Text.Json.Serialization.JsonIgnoreCondition.WhenWritingDefault
         };
 
-        File.WriteAllText(_vmrPath / VmrInfo.SourcesDir / VmrInfo.SourceMappingsFileName,
+        File.WriteAllText(VmrPath / VmrInfo.SourcesDir / VmrInfo.SourceMappingsFileName,
             JsonSerializer.Serialize(sourceMappings, settings));
         
-        await GitOperations.CommitAll(_vmrPath, "Add source mappings");
+        await GitOperations.CommitAll(VmrPath, "Add source mappings");
     }
 }
